requirements.txt
attrs==23.1.0
Automat==22.10.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
constantly==15.1.0
cryptography==41.0.4
Flask==3.0.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
paramiko==3.3.1
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==23.2.0
PyRIC==0.1.6.3
python-nmap==0.7.1
requests==2.31.0
scapy==2.5.0
service-identity==23.1.0
six==1.16.0
soupsieve==2.5
sudo==1.0.0
Twisted==23.8.0
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==3.0.0
zope.interface==6.1
